+++ certlogic/certlogic-kotlin/src/test/kotlin/eu/ehn/dcc/certlogic/PlusTimeTests.kt
@@ -75,4 +75,38 @@ internal class PlusTimeTests {
         check("2004-02-29", -2, TimeUnit.year, "2002-03-01T00:00:00.000Z")
     }
 
+
+    /*
+     * Test with partial dates (in both supported formats), and check whether it coincides with `dccDateOfBirth`.
+     * The tests below have the same partial date strings as for `dccDateOfBirth`, but with a specified offset of 0.
+     */
+
+    @Test
+    fun `works for YYYY the same way it does for dccDateOfBirth`() {
+        check("2004", 0, TimeUnit.hour, "2004-12-31T00:00:00.000Z")
+        check("2021", 0, TimeUnit.hour, "2021-12-31T00:00:00.000Z")
+    }
+
+    @Test
+    fun `works for YYYY-MM the same way it does for dccDateOfBirth`() {
+        check("2004-01", 0, TimeUnit.hour, "2004-01-31T00:00:00.000Z")
+        check("2004-02", 0, TimeUnit.hour, "2004-02-29T00:00:00.000Z")
+        check("2003-02", 0, TimeUnit.hour, "2003-02-28T00:00:00.000Z")
+        check("2004-03", 0, TimeUnit.hour, "2004-03-31T00:00:00.000Z")
+        check("2004-04", 0, TimeUnit.hour, "2004-04-30T00:00:00.000Z")
+        check("2004-05", 0, TimeUnit.hour, "2004-05-31T00:00:00.000Z")
+        check("2004-06", 0, TimeUnit.hour, "2004-06-30T00:00:00.000Z")
+        check("2004-07", 0, TimeUnit.hour, "2004-07-31T00:00:00.000Z")
+        check("2004-08", 0, TimeUnit.hour, "2004-08-31T00:00:00.000Z")
+        check("2004-09", 0, TimeUnit.hour, "2004-09-30T00:00:00.000Z")
+        check("2004-10", 0, TimeUnit.hour, "2004-10-31T00:00:00.000Z")
+        check("2004-11", 0, TimeUnit.hour, "2004-11-30T00:00:00.000Z")
+        check("2004-12", 0, TimeUnit.hour, "2004-12-31T00:00:00.000Z")
+    }
+
+    @Test
+    fun `works for YYYY-MM-DD the same way it does for dccDateOfBirth`() {
+        check("2021-05-04", 0, TimeUnit.hour, "2021-05-04T00:00:00.000Z")
+    }
+
 }